sketch of the prefab
here is what i come up with after 15min of gimp.
My idea is to have the drill on the right and some kind of a tube on the
left which is "suckin" out stuffs that were drilled to then put it in
the store box at the complete left.
The player can then come and pick up what is in the storebox. This way
it also prevent it to be too OP if the storage is limited (could be
configurable).

Do not hesitate to tell me if the artwork is horrible :-)