Widget setup (working for client as well)
So I used my sketch to be able to spawn the prefab. There will just be
to replace the image in the exported folder, the spriter project with
the animations is already set up.

For the storage i went for the following :
When you spawn the mining machine, i spawn a "follower". I put a swap
symbol on the build of the machine. The "follower" is a prefab which is
basically the same than the front door of the storage (we will have to
adjust the offsets when we have the final design).
This way you can interact with the machine (start it, stop it,
whatever...) without being disturbed by the storage part that will be
handled by this "follower" prefab.

I put the storage configurable from 2x2 to 4x4. The widget is scaling
accordingly.

If you don't like the design of the widget, feel free to change it.
Please just keep the exact same dimensions of the image
"exported/mms_widget/mms_bg.png"