A lot of additions...
Changelog:
*Machine component is set. Turning On the machine starts the digging
process, turning it off stops it (functionnal)
*Digging function : Fully functionnal. It digs stuffs according the the
biome it is on and reads the file scripts/mmloottable.lua to get the
probability of each items. It then store the result in the attached
chest.
*Break Down mechanism : Implemented and functionnal. When failing to
dig, there is a currently 15% chance to break down the machine. If the
storage is full but the machine keeps on digging there is a 10% chance
that it breaks down (it damages the mechanism by trying to overfill the
chest). The item is lost in the process.
*I added a new component (mnzmachines.lua) to handle the new action
"MMREPAIR", to repair the machine when it breaks down. currently there
is no wrench implemented so i just use a gears. The MMREPAIR action is
functionnal and implemented.
*There is a chance to "dig" a mob. I added to mmloottable.lua 2 new
etries. I will explain the first one here. CHESTMOBS are the mobs that
can be "dug" and it changes with the biome you are currently on. When a
mob is "dug", it is stored temporarily in a mnzmachines variable
(mnzmachines.surpriseinchest) which is a table which can contain up to 3
elements (you cannot overcrowd your chest...) each with 3 entities, for
a max of 9 mobs to spawn at the same time when the player open the
storage.
*Added the possibility of larger creatures to spawn at the surface next
to the machine. This is also biome dependant and it is handled by the
second new entry in the mmloottable.lua called "ESCAPEMOBS". I tested
it, it's functionnal. To prevent spamming during playtesting I put the
function to be run 4 times per day (to be discussed for the final
value).
*The chance for a mob to be "dug" or to escape are configurable.
*Player cannot put him/herself item in the machine storage. It is just
possible to retrieve items from here.

-> For the final design : For the whole stuff to make sense, the hole
under the machine will have to be big enough so rock lobsters or
bunnyman could pass through it (at least not be completly tiny)